Delete thread

+++ app/Http/Livewire/Thread/Delete.php
@@ -2,12 +2,23 @@
 
 namespace App\Http\Livewire\Thread;
 
+use App\Models\Thread;
+use App\Policies\ThreadPolicy;
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
+use Illuminate\Support\Facades\Auth;
 use Livewire\Component;
 
 class Delete extends Component
 {
+    use AuthorizesRequests;
+
     public $thread;
     public $confirmingThreadDeletion = false;
+
+    /*public function authorize($ability, $arguments = [])
+    {
+        return $this->authorizeForUser(Auth::user(), $ability, $arguments);
+    }*/
 
     public function confirmThreadDeletion()
     {
@@ -18,7 +29,11 @@ class Delete extends Component
 
     public function deleteThread()
     {
+        $this->authorize(ThreadPolicy::DELETE, $this->thread);
+
         $this->thread->delete();
+
+        $this->confirmingThreadDeletion = false;
 
         session()->flash('success', 'Thread deleted!');
